Keep track (and show) title of currently loaded tree
