+++ assets/index.html
@@ -15,6 +15,7 @@
         <noscript>
             <div class="error">JavaScript support is required.</div>
         </noscript>
+        <div id="tree-title">[TreeTitle]</div>
 
         <div class="toolbox">
             <div class="toolbox-element" id="opentree-container">
