+++ assets/toolboxstyle.css
@@ -20,7 +20,7 @@
     display: flex;
     align-items: center;
     text-align: center;
-    font-size: 12px;
+    font-size: 10pt;
 }
 
 #opentree-header, #opentree-file {
