+++ src/app.ts
@@ -1,4 +1,4 @@
-﻿import * as Utils from "./utils";
+﻿import * as DomUtils from "./domutils";
 import * as Sequencer from "./sequencer";
 import * as Tree from "./tree";
 import * as TreeParser from "./treeparser";
@@ -7,8 +7,8 @@ import * as TreeDisplay from "./treedisplay";
 export async function run(): Promise<void> {
     sequencer = Sequencer.createRunner();
 
-    Utils.subscribeToFileInput("opentree-file", enqueueLoadTree);
-    Utils.subscribeToClick("openexample-button", () => enqueueLoadTree("example.tree.json"));
+    DomUtils.subscribeToFileInput("opentree-file", enqueueLoadTree);
+    DomUtils.subscribeToClick("openexample-button", () => enqueueLoadTree("example.tree.json"));
 
     console.log("Started running");
 
@@ -19,22 +19,26 @@ export async function run(): Promise<void> {
 }
 
 let currentTree: Tree.Node | undefined = undefined;
+let currentTitle: string | undefined = undefined;
 let sequencer: Sequencer.SequenceRunner | undefined = undefined;
 
 function enqueueLoadTree(source: string | File): void {
+    let name = typeof source == "string" ? source : source.name;
     sequencer!.enqueue(async () => {
         let result = await TreeParser.load(source);
         if (result.kind == "error")
             alert(`Failed to load. Error: ${result.errorMessage}`);
         else {
-            console.log("Successfully loaded tree");
+            console.log(`Successfully loaded tree: ${name}`);
             // Tree.printNode(result.value, 2);
-            setCurrentTree(result.value);
+            setCurrentTree(result.value, name);
         }
     });
 }
 
-function setCurrentTree(node: Tree.Node): void {
+function setCurrentTree(node: Tree.Node, name: string): void {
     currentTree = node;
+    currentTitle = name;
+    DomUtils.setText("tree-title", name);
     TreeDisplay.setTree(currentTree);
 }
